# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /move/recommend:
    post:
      operationId: getMoveRecommendation
      tags: [entries]
      summary: Get AI mood-based move video recommendation
      description: Recommends the best video from today's rotation based on user's mood using Gemini
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveRecommendationInput"
      responses:
        "200":
          description: AI move recommendation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoveRecommendationResponse"
  /notifications/preferences:
    get:
      operationId: getNotificationPreferences
      tags: [session]
      summary: Get user notification preferences
      description: Returns preferred time slot, enabled status, and whether checked in today
      responses:
        "200":
          description: User notification preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferences"
    post:
      operationId: updateNotificationPreferences
      tags: [session]
      summary: Update user notification preferences
      description: Updates preferred time slot range and enabled status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNotificationPreferencesInput"
      responses:
        "200":
          description: Notification preferences updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferences"
  /notifications/log-engagement:
    post:
      operationId: logNotificationEngagement
      tags: [session]
      summary: Log in-app reminder engagement or dismissal
      description: Records reminder_shown, reminder_engaged, or reminder_dismissed to activity_log
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogNotificationEngagementInput"
      responses:
        "200":
          description: Engagement logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogNotificationEngagementResponse"
...
components:
  schemas:
...
    MoveRecommendationInput:
      type: object
      properties:
        mood:
          type: string
          minLength: 1
          maxLength: 300
      required: [mood]
    MoveRecommendationResponse:
      type: object
      properties:
        mood:
          type: string
        recommendedTitle:
          type: string
        reason:
          type: string
        isCached:
          type: boolean
      required: [mood, recommendedTitle, reason]
    NotificationPreferences:
      type: object
      properties:
        preferredTimeSlot:
          type: string
        enabled:
          type: boolean
        checkedInToday:
          type: boolean
      required: [preferredTimeSlot, enabled, checkedInToday]
    UpdateNotificationPreferencesInput:
      type: object
      properties:
        preferredTimeSlot:
          type: string
        enabled:
          type: boolean
      required: [preferredTimeSlot, enabled]
    LogNotificationEngagementInput:
      type: object
      properties:
        action:
          type: string
          enum: [shown, engaged, dismissed]
        timeSlot:
          type: string
      required: [action]
    LogNotificationEngagementResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
...