# lib/api-zod/src/generated/types/logNotificationEngagementInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LogNotificationEngagementInputAction } from './logNotificationEngagementInputAction';

export interface LogNotificationEngagementInput {
  action: LogNotificationEngagementInputAction;
  timeSlot?: string;
}