# lib/api-zod/src/generated/types/notificationPreferences.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NotificationPreferences {
  preferredTimeSlot: string;
  enabled: boolean;
  checkedInToday: boolean;
}